Terminado (creo) lo del bitmap Falta arreglar los fwrite y fread :(

+++ LFS/Mensajeria/mensajes.h
@@ -28,7 +28,7 @@
 
 typedef struct{
 
-	int key;
+	u_int16_t key;
 	char* value;
 	time_t timestamp;
